# docs/api.yaml
openapi: 3.0.0
info:
  title: 'url-shortener'
  description: 'api to create and manage short urls'
  version: 0.1.0
  contact:
    name: 'Lorenz Dobmeier'
    email: 'lorenz.dobmeier@gmx.de'
servers:
  - url: 'https://example.com'
paths:
...
  /{key}/info:
    get:
      summary: 'get info for key'
      security:
        - oAuth2: []
      parameters:
        - $ref: '#/components/parameters/shorturlkey'
      responses:
        '200':
          $ref: '#/components/responses/info'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notfound'
        '500':
          $ref: '#/components/responses/internalservererror'
  /{key}/stats:
    get:
      summary: 'get statistics for key'
      security:
        - oAuth2: []
      parameters:
        - $ref: '#/components/parameters/shorturlkey'
      responses:
        '200':
          $ref: '#/components/responses/statistics'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notfound'
        '500':
          $ref: '#/components/responses/internalservererror'
components:
  securitySchemes:
  
    oAuth2:
      type: oauth2
      description: 'OAuth2 with implicit grant flow'
      flows:
        implicit:
          authorizationUrl: 'https://login.microsoftonline.com/AZUREAD_APP_TENANTID/oauth2/v2.0/authorize'
          scopes: {}
  responses:
...
    unauthorized:
      description: 'Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    forbidden:
      description: 'Forbidden'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    notfound:
      description: 'Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    internalservererror:
      description: 'Internal Server Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    info:
      description: 'OK'
      content:
        application/json:
          schema:
            type: object
            properties:
              key:
                type: string
                example: 'j209sn2'
              persistent:
                type: boolean
                nullable: false
                example: true
              shorturl:
                type: string
                format: uri
                example: 'https://example.com/j209sn2'
              longurl:
                type: string
                format: uri
                example: 'https://mylongurl.com/service/route/key?param=dkjashbcdnka'
              createdAt:
                type: string
                format: timestamp
                example: "2021-01-01T00:00:00.000Z"
              updatedAt:
                type: string
                format: timestamp
                example: "2021-01-01T00:00:00.000Z"
    statistics:
      description: 'OK'
      content:
        application/json:
          schema:
            type: object
            properties:
              key:
                type: string
                example: 'j209sn2'
              calls:
                type: number
                format: int32
                description: 'no. of requests for specified key'
              uniqueCallers:
                type: number
                format: int32
                description: 'no. of unique callers for specified key'
              callsOfUniqueCallers:
                type: array
                items:
                  type: number
                  format: int32
              history:
                type: array
                items:
                  type: object
                  description: 'dates are only provided, if there were one or more requests that day'
                  properties:
                    date:
                      type: string
                      format: date
                      example: '2021-01-01'
                    calls:
                      type: number
                      format: int32
                      description: 'no. of requests on given date for specified key'
                      example: 52
  parameters:
...
    shorturlkey:
      name: 'key'
      in: path
      required: true
      schema:
        type: string
        minLength: 7
        maxLength: 7
      description: 'key of a short url'
  schemas:
    Error:
      description: 'Error'
      type: object
      properties:
        code:
          type: number
          description: 'http error code'
        message:
          type: string
          description: 'http error message'
        description:
          type: string
          maxLength: 150
          description: 'additional description for error message'
      required:
        - code
        - message
...